+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: use typing (types.od.HighlightStyle) inside `theme.lua`
 - feat: add non-exists highlight from `overrides` option in dev mode (fixed #99)
 - enhance: bright & dark colors updated in terminal themes
+- chore: unlink 'Visual' highlight
 
 ### Changed